requirements.txt
# Gmail API & Authentication
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0

# Data Processing & Excel
pandas==2.1.4
openpyxl==3.11.0

# HTML/Text Parsing
beautifulsoup4==4.12.2
Jinja2==3.1.2

# Configuration
PyYAML==6.0.1

# PDF extraction (optional)
PyPDF2==4.0.1

# Development & Testing
pytest==7.4.3
pytest-mock==3.12.0
black==23.12.0
mypy==1.7.1
